pyproject.toml
[project]
name = "mm_reranker_eval"
version = "0.1.0"
description = "A unified package for evaluating multimodal reranking models"
authors = [
    {name = "Wang Shijian"}
]
readme = "README.md"
requires-python = ">=3.9"
keywords = ["multimodal", "reranking", "evaluation", "retrieval", "cross-modal"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "transformers>=4.47.3",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "pillow>=11.3.0",
    "torchvision>=0.23.0",
    "einops>=0.7.0",
    "timm>=0.9.0",
    "qwen-vl-utils>=0.0.8",
    "requests>=2.28.0",
    "sentence-transformers>=5.1.1",
    "packaging>=20.0",
    "peft>=0.7.0",
]

[project.optional-dependencies]
flash-attention = ["flash-attn>=2.0.0"]
colpali = ["colpali-engine>=0.3.4"]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.build-system]
requires = [
    "setuptools>=45",
    "wheel",
    "setuptools",
    "wheel",
    "packaging",
    "ninja; platform_system != 'Windows'",
]
build-backend = "setuptools.build_meta"

[project.scripts]
mmranker = "mm_reranker_eval.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["mm_reranker_eval*"]

[tool.setuptools]
zip-safe = false
